requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anthropic==0.25.1
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asteroid-filterbanks==0.4.0
attrs==23.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.4
decorator==4.4.2
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
dlib==19.24.4
docopt==0.6.2
docstring-parser==0.15
einops==0.7.0
email_validator==2.1.1
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.110.1
ffmpeg-python==0.2.0
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gpt4all==2.4.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
HyperPyYAML==1.2.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.1
lightning==2.2.1
lightning-utilities==0.11.2
llama-index-core==0.10.28
llama-index-embeddings-openai==0.1.7
llama-index-embeddings-together==0.1.3
llama-index-llms-anthropic==0.1.9
llama-index-llms-openai==0.1.15
llama-index-llms-openai-like==0.1.3
llama-index-llms-together==0.1.3
llama-index-readers-file==0.1.17
llama-index-vector-stores-pinecone==0.1.4
llamaindex-py-client==0.1.18
llvmlite==0.42.0
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
openai==1.17.1
openai-whisper==20231117
opencv-python==4.9.0.80
optuna==3.6.1
orjson==3.10.0
packaging==24.0
pandas==2.2.1
pillow==10.3.0
pinecone-client==3.2.2
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
proglog==0.1.10
protobuf==5.26.1
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydantic==2.7.0
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
Pygments==2.17.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.2.1
pytorch-metric-learning==2.5.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
semver==3.0.2
sentencepiece==0.2.0
serpapi==0.1.5
setuptools==69.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.29
starlette==0.37.2
striprtf==0.0.26
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.2
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.43.0
whisper==1.1.10
wrapt==1.16.0
yarl==1.9.4
